requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
argon2-cffi==20.1.0
async_generator==1.10
attrs==20.3.0
backcall==0.2.0
blas==1.0
bleach==3.3.0
ca-certificates==2020.10.14
certifi==2020.6.20
cffi==1.14.5
charset-normalizer==2.0.2
click==7.1.2
cloudpickle==1.6.0
colorama==0.4.4
cudatoolkit==10.2.89
cycler==0.10.0
cytoolz==0.11.0
dask-core==2021.4.0
decorator==5.0.6
defusedxml==0.7.1
entrypoints==0.3
flask==1.1.2
freetype==2.10.4
fsspec==0.9.0
gunicorn==20.1.0
hdf5==1.8.20
heroku==0.1.4
icc_rt==2019.0.0
icu==58.2
idna==3.2
imageio==2.9.0
importlib-metadata==3.10.0
intel-openmp==2021.2.0
ipykernel==5.3.4
ipython==7.22.0
ipython_genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==1.1.0
jedi==0.17.0
jinja2==2.11.3
joblib==1.0.1
jpeg==9b
jsonschema==3.2.0
jupyter==1.0.0
jupyter_client==6.1.12
jupyter_console==6.4.0
jupyter_core==4.7.1
jupyterlab_pygments==0.1.2
jupyterlab_widgets==1.0.0
kiwisolver==1.3.1
libopencv==3.4.2
libpng==1.6.37
libprotobuf==3.16.0
libsodium==1.0.18
libtiff==4.1.0
libuv==1.40.0
locket==0.2.1
lz4-c==1.9.3
m2w64-gcc-libgfortran==5.3.0
m2w64-gcc-libs==5.3.0
m2w64-gcc-libs-core==5.3.0
m2w64-gmp==6.1.0
m2w64-libwinpthread-git==5.0.0.4634.697f757
markupsafe==1.1.1
matplotlib==3.3.4
matplotlib-base==3.3.4
mistune==0.8.4
mkl==2021.2.0
mkl-service==2.3.0
mkl_fft==1.3.0
mkl_random==1.2.1
msys2-conda-epoch==20160418
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5
ninja==1.10.2
notebook==6.3.0
numpy==1.20.1
numpy-base==1.20.1
olefile==0.46
opencv==3.4.2
openssl==1.1.1k
packaging==20.9
pandas==1.2.4
pandoc==2.12
pandocfilters==1.4.3
parso==0.8.2
partd==1.2.0
pickleshare==0.7.5
pillow==8.2.0
pip==21.0.1
prometheus_client==0.10.1
prompt-toolkit==3.0.17
protobuf==3.16.0
py-opencv==3.4.2
pycparser==2.20
pygments==2.8.1
pyparsing==2.4.7
pyqt==5.9.2
pyrsistent==0.17.3
python==3.7.10
python-dateutil==1.5
python_abi==3.7
pytorch==1.8.0
pytz==2021.1
pywavelets==1.1.1
pywin32==227
pywinpty==0.5.7
pyyaml==5.4.1
pyzmq==20.0.0
qt==5.9.7
qtconsole==5.0.3
qtpy==1.9.0
requests==2.26.0
rstr==2.2.6
scikit-image==0.18.1
scikit-learn==0.24.1
scipy==1.6.2
send2trash==1.5.0
setuptools==52.0.0
sip==4.19.8
six==1.15.0
sqlite==3.35.4
tensorboardx==2.2
terminado==0.9.4
testpath==0.4.4
threadpoolctl==2.1.0
tifffile==2020.10.1
tk==8.6.10
toolz==0.11.1
torchaudio==0.8.0
torchvision==0.9.0
tornado==6.1
tqdm==4.59.0
traitlets==5.0.5
typing_extensions==3.7.4.3
urllib3==1.26.6
vc==14.2
vs2015_runtime==14.27.29016
wcwidth==0.2.5
webencodings==0.5.1
werkzeug==1.0.1
wheel==0.36.2
widgetsnbextension==3.5.1
wincertstore==0.2
winpty==0.4.3
xz==5.2.5
yaml==0.2.5
zeromq==4.3.3
zipp==3.4.1
zlib==1.2.11
zstd==1.4.9
